Trying to make cover to draw
And fall back to default cover if that can’t be found.

NOTE: Cover doesn’t draw, regardless default cover or not.
NOTE2: Default cover added to gitignore as I test with copyrighted work. You’re welcome to commit a fallback cover if you hold copyright and want to grant use for this repo, remember to also add this information to the readme if you do so!